Isolate Rediscala configured target tests
